Prolong lifetime of SSHFS object as long as all child SSHFile objects
This fixes the issue with downloading large files that are located on SFTP storage.

The filesystem object was destroyed because no one was referencing it, while the SSHFile object was still alive (opened by Apache file streamer).
As a result, a downloaded file was cut off (sometimes at 1MB size, sometimes bigger)